backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
paho-mqtt==1.6.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
pymongo
